FACES-2696 Utilize Portlet 3.0 dynamic resource dependency feature (backport relevant changes)
