+++ tck/bridge-tck-common/src/main/java/com/liferay/faces/bridge/tck/common/util/faces/context/TCK_FacesContextFactoryImpl.java
@@ -15,8 +15,6 @@
  */
 package com.liferay.faces.bridge.tck.common.util.faces.context;
 
-import com.liferay.faces.bridge.tck.common.util.BridgeTCKUtil;
-
 import javax.faces.FacesException;
 import javax.faces.FactoryFinder;
 import javax.faces.context.FacesContext;
@@ -147,7 +145,7 @@ public class TCK_FacesContextFactoryImpl extends FacesContextFactory {
 		}
 		else {
 
-			if ((phase != null) && (BridgeTCKUtil.isHeaderOrRenderPhase(phase))) {
+			if (Bridge.PortletPhase.RENDER_PHASE.equals(phase)) {
 				request.setAttribute("javax.portlet.faces.tck.verifyPortletPhaseDuringRenderPass",
 					"Phase attribute correctly set during render.");
 			}
