+++ tck/bridge-tck-harness/src/main/java/com/liferay/faces/bridge/tck/harness/TckTestCase.java
@@ -180,7 +180,7 @@ public class TckTestCase extends IntegrationTesterBase {
 		}
 
 		Enumeration<?> tests = testProps.propertyNames();
-		Pattern extractTestNamePattern = Pattern.compile("chapter[0-9_]+Tests-([a-zA-Z]+)-portlet");
+		Pattern extractTestNamePattern = Pattern.compile("chapter[0-9_]+Tests-([_a-zA-Z0-9]+)-portlet");
 		Pattern testFilterPattern = null;
 		String testFilter = TestUtil.getSystemPropertyOrDefault("integration.test.filter", null);
 
@@ -501,7 +501,6 @@ public class TckTestCase extends IntegrationTesterBase {
 
 				browser.switchTo().frame("tck-iframe");
 				browser.waitForElementVisible("//body");
-
 				testPage(browser, false);
 			}
 			else {
