+++ tck/bridge-tck-main-portlet/src/main/java/com/liferay/faces/bridge/tck/tests/chapter_6/section_6_1_3_1/Tests.java
@@ -52,7 +52,6 @@ import javax.xml.namespace.QName;
 import com.liferay.faces.bridge.tck.annotation.BridgeTest;
 import com.liferay.faces.bridge.tck.beans.TestRunnerBean;
 import com.liferay.faces.bridge.tck.common.Constants;
-import com.liferay.faces.bridge.tck.common.util.BridgeTCKUtil;
 import com.liferay.faces.bridge.tck.common.util.HTTPUtils;
 
 
@@ -1315,7 +1314,8 @@ public class Tests extends Object {
 		// This tests that we can encode a new mode in an actionURL
 		// done by navigation rule.
 		FacesContext ctx = FacesContext.getCurrentInstance();
-		if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+
+		if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 			Map<String, String> requestParameterMap = ctx.getExternalContext().getRequestParameterMap();
 
 			String testName = testRunner.getTestName();
@@ -2635,7 +2635,7 @@ public class Tests extends Object {
 
 			return Constants.TEST_SUCCESS; // action Navigation result
 		}
-		else if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+		else if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 
 			testRunner.setTestComplete(true);
 
@@ -2784,7 +2784,7 @@ public class Tests extends Object {
 
 			return Constants.TEST_SUCCESS; // action Navigation result
 		}
-		else if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+		else if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 
 			testRunner.setTestComplete(true);
 
@@ -3043,7 +3043,7 @@ public class Tests extends Object {
 
 			return Constants.TEST_SUCCESS; // action Navigation result
 		}
-		else if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+		else if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 
 			testRunner.setTestComplete(true);
 
@@ -3594,7 +3594,7 @@ public class Tests extends Object {
 
 			return Constants.TEST_SUCCESS; // action Navigation result
 		}
-		else if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+		else if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 
 			testRunner.setTestComplete(true);
 
@@ -4839,7 +4839,7 @@ public class Tests extends Object {
 
 			return Constants.TEST_SUCCESS; // action Navigation result
 		}
-		else if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+		else if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 
 			testRunner.setTestComplete(true);
 
@@ -4986,7 +4986,7 @@ public class Tests extends Object {
 
 			return Constants.TEST_SUCCESS; // action Navigation result
 		}
-		else if (BridgeTCKUtil.isHeaderOrRenderPhase(ctx)) {
+		else if (Bridge.PortletPhase.RENDER_PHASE.equals(BridgeUtil.getPortletRequestPhase())) {
 
 			testRunner.setTestComplete(true);
