+++ tck/bridge-tck-main-portlet/src/main/java/com/liferay/faces/bridge/tck/tests/chapter_6/section_6_5_1/Tests.java
@@ -159,8 +159,8 @@ public class Tests {
 		}
 
 		// HeaderRequest / RenderRequest
-		else
-		{
+		else {
+
 			try {
 
 				// JSF Implicit Objects
